refactor: remove createState function and integrate its functionality directly into createStore for improved state management
