+++ src/core/store.js
@@ -1,8 +1,24 @@
-import createState from "./state.js";
+import createObserver from "./observer.js";
 
-export default function createStore(initialState) {
-  return createState(initialState);
+export default function createStore(initialState = {}) {
+  let state = { ...initialState };
+
+  // 옵저버로 리스너 관리
+  const { subscribe, notify } = createObserver();
+
+  function setState(partial) {
+    // 변경 사항이 없으면 무시
+    if (!partial || (typeof partial === "object" && Object.keys(partial).length === 0)) return;
+    state = { ...state, ...partial };
+    notify(state);
+  }
+
+  /** 현재 상태를 반환 */
+  const getState = () => state;
+
+  return {
+    subscribe,
+    setState,
+    getState,
+  };
 }
-
-// 선택적으로 새 API를 동일 파일에서 재노출하여 점진적 마이그레이션 지원
-export { createState };
